Move the addLocalField method to abstract class
